Move NonNull into its own file
