+++ polygraph/types/tests/test_type_definitions.py
@@ -2,8 +2,8 @@ from unittest import TestCase
 
 from polygraph.types.api import typedef
 from polygraph.types.list import List
+from polygraph.types.nonnull import NonNull
 from polygraph.types.scalar import Float, Int, String
-from polygraph.types.type_builder import NonNull
 from polygraph.types.union import Union
